requirements.txt
# Core
Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9

# Environment
python-decouple==3.8

# Auth & Security
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# Utilities
Pillow==10.3.0
django-filter==24.2

# Production server
gunicorn==22.0.0

# Dev / testing
pytest==8.2.0
pytest-django==4.8.0
factory-boy==3.3.0
